requirements/local.txt
-r ./base.txt

Werkzeug==0.16.1 # https://github.com/pallets/werkzeug
ipdb==0.12.3  # https://github.com/gotcha/ipdb
Sphinx==2.3.1  # https://github.com/sphinx-doc/sphinx
psycopg2-binary==2.8.4  # https://github.com/psycopg/psycopg2
django-heroku

# Testing
# ------------------------------------------------------------------------------
mypy==0.761  # https://github.com/python/mypy
django-stubs==1.4.0  # https://github.com/typeddjango/django-stubs
pytest==5.3.4  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.2  # https://github.com/Frozenball/pytest-sugar

# Code quality
# ------------------------------------------------------------------------------
flake8==3.7.9  # https://github.com/PyCQA/flake8
coverage==5.0.3  # https://github.com/nedbat/coveragepy
black==19.10b0  # https://github.com/ambv/black
pylint-django==2.0.13  # https://github.com/PyCQA/pylint-django
pylint-celery==0.3  # https://github.com/PyCQA/pylint-celery
pre-commit==1.21.0  # https://github.com/pre-commit/pre-commit
yapf # https://github.com/google/yapf

# Django
# ------------------------------------------------------------------------------
factory-boy==2.12.0  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==2.1  # https://github.com/jazzband/django-debug-toolbar
django-extensions==2.2.6  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==1.8.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==3.8.0  # https://github.com/pytest-dev/pytest-django
slackclient
slackeventsapi
